// src/models/sale.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bespoked Bikes - API
 * BeSpoked Bikes is looking to create a sales tracking application.  BeSpoked is a high-end bicycle shop and each salesperson gets a commission for each bike they sell.  They are introducing a new quarterly bonus based on sales as an incentive.
 *
 * OpenAPI spec version: 1.0.11
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Sale
 */
export interface Sale {
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    product_id?: number;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    customer_id?: number;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    salesperson_id?: number;
    /**
     * 
     * @type {string}
     * @memberof Sale
     */
    sale_date?: string;
    /**
     * 
     * @type {number}
     * @memberof Sale
     */
    quantity?: number;
}